requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-core==1.4.1
django-cors-headers==3.5.0
django-currentuser==0.5.2
django-environ==0.4.5
django-extensions==3.0.9
django-filter==2.4.0
django-modeltranslation==0.16.1
django-schema-graph==1.2.0
djangorestframework==3.12.2
#djangorestframework==3.9
djangorestframework-simplejwt==4.6.0
drf-yasg2==1.19.4
environ==1.0
idna==2.10
importlib-metadata==3.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
rest-framework-generic-relations==2.0.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
